# swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Dapolac-backend",
    "description" : "API REST Proyecto Final de Ingenieria",
    "contact" : {
      "email" : "gustavowroldan@gmail.com"
    },
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "2.0.0"
  },
  "servers" : [ {
    "url" : "https://daprolac.herokuapp.com/api/v1"
  } ],
...
  "paths" : {
...
    "/tareas" : {
      "get" : {
        "tags" : [ "administradores" ],
        "summary" : "obtener tareas",
        "description" : "Regresa todas las tareas\n",
        "operationId" : "obtenerTareas",
        "parameters" : [ {
          "name" : "eager",
          "in" : "query",
          "description" : "carga todas las asociaciones",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "obtiene las tareas segun el criterio aplicado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_6"
                }
              }
            }
          },
          "500" : {
            "description" : "Error interno del servidor",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "administradores" ],
        "summary" : "agregar tarea",
        "description" : "agregar una nueva tarea al sistema",
        "operationId" : "crearTarea",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_4"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "proceso creado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_201_1"
                }
              }
            }
          },
          "400" : {
            "description" : "entrada invalida u objeto invalido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "500" : {
            "description" : "error interno del servidor",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tareas/{idTarea}" : {
      "get" : {
        "tags" : [ "administradores" ],
        "summary" : "obtener tarea por id",
        "description" : "Regresa una tarea particular\n",
        "operationId" : "obtenerTarea",
        "parameters" : [ {
          "name" : "idTarea",
          "in" : "path",
          "description" : "Identificador unico de la tarea",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "eager",
          "in" : "query",
          "description" : "carga todas las asociaciones",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operacion exitosa",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_6"
                }
              }
            }
          },
          "500" : {
            "description" : "Error interno del servidor",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "administradores" ],
        "summary" : "actualizar tarea existente",
        "description" : "actualizar una tarea existente en el sistema",
        "operationId" : "actualizarTarea",
        "parameters" : [ {
          "name" : "idTarea",
          "in" : "path",
          "description" : "Identificador unico de la tarea",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_5"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "proceso actualizado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_7"
                }
              }
            }
          },
          "400" : {
            "description" : "entrada invalida u objeto invalido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "500" : {
            "description" : "error interno del servidor",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "administradores" ],
        "summary" : "eliminar tarea existente",
        "description" : "eliminar una tarea existente del sistema",
        "operationId" : "eliminarTarea",
        "parameters" : [ {
          "name" : "idTarea",
          "in" : "path",
          "description" : "Identificador unico de la tarea",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "tarea eliminada",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_8"
                }
              }
            }
          },
          "500" : {
            "description" : "error interno del servidor",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Error" : {
        "required" : [ "error", "payload", "success" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean",
            "example" : false
          },
          "payload" : {
            "type" : "object",
            "nullable" : true
          },
          "error" : {
            "$ref" : "#/components/schemas/Error_error"
          }
        }
      },
...
      "Tarea" : {
        "required" : [ "id", "nombre" ],
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 4
          },
          "nombre" : {
            "type" : "string",
            "example" : "Tarea 4"
          },
          "observaciones" : {
            "type" : "string",
            "example" : "Esta tarea especifica..."
          }
        }
      },
...
      "inline_response_201_1" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "$ref" : "#/components/schemas/inline_response_201_1_payload"
          },
          "error" : {
            "type" : "object",
            "nullable" : true
          }
        }
      },
...
      "inline_response_200_6" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Tarea"
            }
          },
          "error" : {
            "type" : "object",
            "nullable" : true
          }
        }
      },
      "tareas_proceso_tareaAntecesora" : {
        "required" : [ "diasAntecesora", "horasAntecesora", "idTareaAntecesora", "minutosAntecesora" ],
        "type" : "object",
        "properties" : {
          "idTareaAntecesora" : {
            "type" : "integer",
            "description" : "tarea antecesora que se debe cumplir \npara comenzar la actual\n"
          },
          "diasAntecesora" : {
            "type" : "integer",
            "description" : "dias que deben pasar para comenzar esta \ntarea luego de finalizada la antecesora.\n"
          },
          "horasAntecesora" : {
            "type" : "integer",
            "description" : "horas que deben pasar para comenzar esta \ntarea luego de finalizada la antecesora.\n"
          },
          "minutosAntecesora" : {
            "type" : "integer",
            "description" : "minutos que deben pasar para comenzar esta \ntarea luego de finalizada la antecesora.\n"
          }
        }
      },
      "tareas_proceso" : {
        "required" : [ "idProceso" ],
        "type" : "object",
        "properties" : {
          "idProceso" : {
            "type" : "integer",
            "description" : "identificador del proceso al que \nse quiere asociar la tarea\n"
          },
          "tareaAntecesora" : {
            "$ref" : "#/components/schemas/tareas_proceso_tareaAntecesora"
          }
        }
      },
      "body_4" : {
        "required" : [ "nombre" ],
        "type" : "object",
        "properties" : {
          "nombre" : {
            "type" : "string",
            "example" : "Seleccion de la leche"
          },
          "observaciones" : {
            "type" : "string",
            "description" : "aclaraciones sobre la tarea"
          },
          "proceso" : {
            "$ref" : "#/components/schemas/tareas_proceso"
          }
        }
      },
      "body_5" : {
        "required" : [ "producto" ],
        "type" : "object",
        "properties" : {
          "producto" : {
            "type" : "string"
          }
        }
      },
      "inline_response_200_7" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "$ref" : "#/components/schemas/inline_response_200_7_payload"
          },
          "error" : {
            "type" : "object",
            "nullable" : true
          }
        }
      },
      "inline_response_200_8" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "payload" : {
            "$ref" : "#/components/schemas/inline_response_200_8_payload"
          },
          "error" : {
            "type" : "object",
            "nullable" : true
          }
        }
      },
      "Error_error" : {
        "type" : "object",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "example" : 400
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "inline_response_201_1_payload" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "identificador del proceso creado.",
            "example" : 391
          }
        }
      },
...
      "inline_response_200_7_payload" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "identificador de la tarea actualizada.",
            "example" : 45
          }
        }
      },
      "inline_response_200_8_payload" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "identificador de la tarea eliminada.",
            "example" : 111
          }
        }
      }
    }
  }
}